+++ docs/hooks.py
@@ -1,11 +1,8 @@
-import os
-import shutil
 import subprocess
 
 
 def on_post_build(config, **kwargs):
     site_dir = config["site_dir"]
     subprocess.run(
-        ["pdoc", "-t", "./docs/assets/css", "./dbterd/api/base.py", "-o", "./target/pdoc"]
+        ["pdoc", "-t", "docs/assets/css", "dbterd/api/", "-o", f"{site_dir}/api-docs"]
     )
-    shutil.copytree("target/pdoc", os.path.join(site_dir, "api-docs"))
